+++ src/impls/$consume/async-consume.js
@@ -8,32 +8,14 @@
 
 import { asyncIterableCurry } from '../../internal/async-iterable.js';
 
-let warnedCallbackDeprecation = false;
-
-const warnCallbackDeprecation = () => {
-  if (!warnedCallbackDeprecation) {
-    console.warn(
-      `\`${'asyncConsume'}(callback, iterable)\` is deprecated and will be removed in iter-tools@8. ` +
-        `Instead use ${'asyncForEach'}(callback, iterable)`,
-    );
-    warnedCallbackDeprecation = true;
-  }
-};
-
-export async function __asyncConsume(iterable, callback = () => {}) {
-  let c = 0;
-  for await (const value of iterable) {
-    await callback(value, c++);
-  }
+export async function __asyncConsume(iterable) {
+  /* eslint-disable */
+  // prettier-ignore
+  for await (const _ of iterable)
+    {}
+  /* eslint-enable */
 }
 
 export const asyncConsume = /*#__PURE__*/ asyncIterableCurry(__asyncConsume, {
   reduces: true,
-  minArgs: 0,
-  maxArgs: 1,
-  validateArgs(args) {
-    if (typeof args[1] === 'function') {
-      warnCallbackDeprecation();
-    }
-  },
 });